+++ packages/mantine/src/components/table/__tests__/TablePredicate.spec.tsx
@@ -11,6 +11,11 @@ const columns: Array<ColumnDef<RowData>> = [columnHelper.accessor('name', {enabl
 
 describe('Table.Predicate', () => {
     it('goes back to the first page when changing the predicate', async () => {
+        if (!HTMLElement.prototype.scrollIntoView) {
+            HTMLElement.prototype.scrollIntoView = () => {
+                vi.fn();
+            };
+        }
         const user = userEvent.setup();
         const data = [{name: 'fruit'}, {name: 'vegetable'}];
         const Fixture = () => {
@@ -39,5 +44,69 @@ describe('Table.Predicate', () => {
         await user.click(screen.getByRole('textbox', {name: 'Rank'}));
         await user.click(screen.getByRole('option', {name: 'First'}));
         expect(screen.getByRole('button', {name: '1', current: 'page'})).toBeVisible();
+    });
+
+    describe('when url sync is activated', () => {
+        afterEach(() => {
+            window.history.replaceState(null, '', '/');
+        });
+
+        it('sets the current predicate value in the url using the predicate id as key', async () => {
+            const user = userEvent.setup();
+            const data = [{name: 'fruit'}, {name: 'vegetable'}];
+            const Fixture = () => {
+                const store = useTable<RowData>({
+                    initialState: {predicates: {rank: 'ALL'}},
+                    syncWithUrl: true,
+                });
+                return (
+                    <Table store={store} data={data} columns={columns}>
+                        <Table.Header>
+                            <Table.Predicate
+                                id="rank"
+                                label="Rank"
+                                data={[
+                                    {value: 'ALL', label: 'All'},
+                                    {value: 'first', label: 'First'},
+                                    {value: 'second', label: 'Second'},
+                                ]}
+                            />
+                        </Table.Header>
+                    </Table>
+                );
+            };
+            render(<Fixture />);
+            await user.click(screen.getByRole('textbox', {name: 'Rank'}));
+            await user.click(screen.getByRole('option', {name: 'First'}));
+            expect(window.location.search).toBe('?rank=first');
+        });
+
+        it('determines the initial predicate value from the url', async () => {
+            window.history.replaceState(null, '', '?rank=second');
+            const data = [{name: 'fruit'}, {name: 'vegetable'}];
+            const Fixture = () => {
+                const store = useTable<RowData>({
+                    initialState: {predicates: {rank: 'ALL'}},
+                    syncWithUrl: true,
+                });
+                return (
+                    <Table store={store} data={data} columns={columns}>
+                        <Table.Header>
+                            <Table.Predicate
+                                id="rank"
+                                label="Rank"
+                                data={[
+                                    {value: 'ALL', label: 'All'},
+                                    {value: 'first', label: 'First'},
+                                    {value: 'second', label: 'Second'},
+                                ]}
+                            />
+                        </Table.Header>
+                    </Table>
+                );
+            };
+            render(<Fixture />);
+            expect(screen.getByRole('textbox', {name: 'Rank'})).toHaveValue('Second');
+        });
     });
 });