+++ CHANGELOG.md
@@ -1,13 +1,23 @@
-## <small>55.3.4-next.0 (2025-01-02)</small>
+## 55.5.0 (2025-01-08)
 
-* feat(mantine): remove a lot of custom style to get a vanilla style (#3986) 354db73, closes #3986
+* feat(mantine): export useUrlSyncedState and support hash router [ADUI-10457] (#3992) a10fbd0, closes #3992
+* chore(deps): lock file maintenance (#3989) f375705, closes #3989
+
+
+
+## 55.4.0 (2025-01-03)
+
+* fix(deps): update dnd-kit monorepo (#3983) 1edb5b3, closes #3983
+* fix(table-date-picker): date range picker end date at end of day (#3988) 1b33e19, closes #3988
 * chore(deps): lock file maintenance (#3978) 116e5c5, closes #3978
 * chore(deps): lock file maintenance (#3984) 34f8056, closes #3984
+* chore(deps): lock file maintenance (#3987) 694141e, closes #3987
 * chore(deps): update all non-major dependencies (#3975) bc03a26, closes #3975
 * chore(deps): update all non-major dependencies (#3976) 51ce304, closes #3976
 * chore(deps): update all non-major dependencies (#3979) 3f07f5f, closes #3979
 * chore(deps): update all non-major dependencies (#3980) c35f070, closes #3980
 * chore(deps): update dependency stylelint-config-standard-scss to v14 (#3977) 9db7974, closes #3977
+* feat(mantine): add syncWithUrl option to table hook (#3985) 71740c9, closes #3985